users_service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
asyncpg>=0.24.0
tenacity>=8.0.0
pydantic>=1.8.0
aioredis>=2.0.1
bcrypt>=4.3.0
python-jose[cryptography]>=3.3.0
email_validator>=2.2.0